+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/project_catalog/_connect.py
@@ -19,9 +19,9 @@ class Connect(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-05-01-preview",
+        "version": "2024-10-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/catalogs/{}/connect", "2024-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/catalogs/{}/connect", "2024-10-01-preview"],
         ]
     }
 
@@ -150,12 +150,12 @@ class Connect(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-05-01-preview",
+                    "api-version", "2024-10-01-preview",
                     required=True,
                 ),
             }
             return parameters
-        
+
         def on_200(self, session):
             pass