+++ src/aks-preview/HISTORY.rst
@@ -13,6 +13,10 @@ To release a new version, please select a new version number (usually plus 1 to 
 +++++++
 * `az aks create/update`: Update advanced container networking service (acns) with 2024-09-02-preview API version enablement.
 
+Pending
++++++++
+* Vendor new SDK and bump API version to 2024-09-02-preview.
+
 13.0.0b1
 +++++++
 * [BREAKING CHANGE]: `az aks trustedaccess rolebinding create`: Remove deprecated `-r` and `-s` options.